+++ src/app/search/search.controller.js
@@ -5,7 +5,7 @@
         .controller('SearchController', SearchController);
 
     /** @ngInject */
-    function SearchController ($analytics, $filter, $interval, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, CACHE_REFRESH_TIMEOUT, CACHE_TIMEOUT, RELOAD_TIMEOUT, commonService, utilService) {
+    function SearchController ($analytics, $filter, $interval, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, CACHE_REFRESH_TIMEOUT, CACHE_TIMEOUT, RELOAD_TIMEOUT, SPLIT_PRIMARY, commonService, utilService) {
         var vm = this;
 
         vm.browseAll = browseAll;
@@ -39,9 +39,6 @@
         ////////////////////////////////////////////////////////////////////
 
         function activate () {
-            if ($localStorage.clearResults) {
-                vm.clear();
-            }
             $scope.$on('ClearResults', function () {
                 vm.clear();
             });
@@ -55,10 +52,14 @@
             vm.isPreLoading = true;
             vm.restoreStateHs = [];
             vm.showRetiredHs = [];
+            vm.SPLIT_PRIMARY = SPLIT_PRIMARY;
 
             manageStorage();
             populateSearchOptions();
             restoreResults();
+            if ($localStorage.clearResults) {
+                vm.clear();
+            }
             vm.loadResults();
             setTimestamp();
         }